Quality of life fixes
Changed sneak to toggle, changed it to c key to avoid issues during gameplay.
Added two new behaviour scripts for NPCs, and modified point wander behaviour.
Modified the barrel hiding mechanic so it puts you in the barrel you highlighted instead of the nearest one, and the highlight only shows up when you're in range to interact with it.