Behaviour Tree improved
Refactored constructors so I don't have to keep making lists for no reason.
Added a line to pull out of timers. Need to test more to see if it works properly. It might not.
Fixed the turn to functionality when the enemies pause when first spotting the player. It actually works now.